[Refactor] #189 - NotificationData/Domain의 NovelNotificationSetting을 NovelNotification으로 통합
작품 알림 구독 목록(#188)과 작품별 알림 on/off 설정(#189)이 별개 하위 영역(NovelNotificationSetting/)으로 갈려 있던 걸, 둘 다 '작품 알림'이라는 같은 개념을 다른 각도(여러 작품 목록 vs 단일 작품 상태)로 다룰 뿐이라 NovelNotification/ 하나로 합쳤다.

- Domain: NovelNotificationRepository 프로토콜에 loadNotificationSetting/updateNotificationSetting 추가, NovelNotificationSetting 엔티티·두 UseCase를 NovelNotification/으로 이동, MockNovelNotificationRepository에 설정 조회/변경 트래킹 병합
- Data: NovelNotificationEndpoint/Service/Mapper/Logger/Repository에 설정 조회(GET)·변경(PUT) 케이스 병합, DTO 이동. NotificationDataFactory는 makeNovelNotificationRepository 하나로 통일(makeNovelNotificationSettingRepository 제거)
- Feature: NovelDetailFeatureDemoApp의 Repository 생성 호출부를 통합 Factory 메서드로 갱신(UseCase 시그니처는 불변)
- NotificationDomain/NotificationData CLAUDE.md를 3영역(기존 4영역) 구조로 갱신
