+++ Projects/Data/NotificationData/Sources/NovelNotification/DTO/Request/NovelNotificationSettingRequest.swift
