+++ Projects/Domain/NotificationDomain/Sources/NovelNotification/Repository/NovelNotificationRepository.swift
@@ -20,4 +20,13 @@ public protocol NovelNotificationRepository: Sendable {
         type: NovelNotificationType,
         novelIDs: [NovelID]
     ) async throws(RepositoryError)
+
+    /// 작품 하나의 현재 완결/휴재복귀 알림 설정을 조회한다(#189, 작품 상세 종 아이콘 시트).
+    func loadNotificationSetting(novelID: NovelID) async throws(RepositoryError) -> NovelNotificationSetting
+
+    /// PUT은 멱등 — `setting`의 두 값을 항상 함께 보낸다(서버가 부분 갱신을 지원하지 않음).
+    func updateNotificationSetting(
+        novelID: NovelID,
+        setting: NovelNotificationSetting
+    ) async throws(RepositoryError)
 }
